fixes: a lot of individual fixes - fixed basic machines info being inconsistent - cleaned Formatter - fixed generator production value display - changed power - tier - show heat value when above 0 only - added missing item names - added Personal blocks info provider - removed unused field accessor - register wrenchable provider last

+++ src/main/java/reforged/mods/blockhelper/addons/integrations/ic2/GeneratorInfoProvider.java
@@ -19,7 +19,7 @@ public class GeneratorInfoProvider extends InfoProvider {
         if (blockEntity instanceof TileEntityBaseGenerator) {
             TileEntityBaseGenerator gen = (TileEntityBaseGenerator) blockEntity;
             text(helper, tier(1));
-            double production = 0;
+            int production = 0;
             if (gen instanceof TileEntityGenerator || gen instanceof TileEntityGeoGenerator) {
                 if (gen.isConverting()) {
                     production = gen.production; // return production only when converting fuel
